eliminar NavBar y se cambio el nombre del componente Footer
