--- frontend/src/components/common/NabvarLayout.vue
@@ -1,13 +1,0 @@
-<template>
-
-
-</template>
-
-
-
-<style>
-
-
-
-
-</style>
